add run queries to threads/runs, remove previous impl
