+++ apps/backend/convex/_generated/api.d.ts
@@ -16,7 +16,6 @@ import type {
 import type * as action_evaluateMessageUrls from "../action/evaluateMessageUrls.js";
 import type * as action_generateThreadTitle from "../action/generateThreadTitle.js";
 import type * as action_ingestImageUrl from "../action/ingestImageUrl.js";
-import type * as action_run from "../action/run.js";
 import type * as action_textToImage from "../action/textToImage.js";
 import type * as constants from "../constants.js";
 import type * as crons from "../crons.js";
@@ -65,11 +64,6 @@ import type * as entities_patterns_db from "../entities/patterns/db.js";
 import type * as entities_patterns_ent from "../entities/patterns/ent.js";
 import type * as entities_patterns_public from "../entities/patterns/public.js";
 import type * as entities_patterns_validators from "../entities/patterns/validators.js";
-import type * as entities_runs_db from "../entities/runs/db.js";
-import type * as entities_runs_ent from "../entities/runs/ent.js";
-import type * as entities_runs_internal from "../entities/runs/internal.js";
-import type * as entities_runs_public from "../entities/runs/public.js";
-import type * as entities_runs_validators from "../entities/runs/validators.js";
 import type * as entities_shared from "../entities/shared.js";
 import type * as entities_texts_ent from "../entities/texts/ent.js";
 import type * as entities_texts_internal from "../entities/texts/internal.js";
@@ -124,7 +118,6 @@ declare const fullApi: ApiFromModules<{
   "action/evaluateMessageUrls": typeof action_evaluateMessageUrls;
   "action/generateThreadTitle": typeof action_generateThreadTitle;
   "action/ingestImageUrl": typeof action_ingestImageUrl;
-  "action/run": typeof action_run;
   "action/textToImage": typeof action_textToImage;
   constants: typeof constants;
   crons: typeof crons;
@@ -173,11 +166,6 @@ declare const fullApi: ApiFromModules<{
   "entities/patterns/ent": typeof entities_patterns_ent;
   "entities/patterns/public": typeof entities_patterns_public;
   "entities/patterns/validators": typeof entities_patterns_validators;
-  "entities/runs/db": typeof entities_runs_db;
-  "entities/runs/ent": typeof entities_runs_ent;
-  "entities/runs/internal": typeof entities_runs_internal;
-  "entities/runs/public": typeof entities_runs_public;
-  "entities/runs/validators": typeof entities_runs_validators;
   "entities/shared": typeof entities_shared;
   "entities/texts/ent": typeof entities_texts_ent;
   "entities/texts/internal": typeof entities_texts_internal;
